don't print the graph schema in logs
